Adding tracker id's and simplifying configuration.

+++ Sources/iOSBasics/SyncServer/SyncServer+Delegate.swift
@@ -30,24 +30,30 @@ extension SyncServer: ServerAPIDelegate {
     
     func uploadCompleted(_ delegated: AnyObject, result: Swift.Result<UploadFileResult, Error>) {
 
+        logger.debug("uploadCompleted: \(result)")
+
         switch result {
         case .failure(let error):
             delegate.error(self, error: error)
             
         case .success(let uploadFileResult):
             var fileUUID: UUID!
+            var uploadObjectTrackerId: Int64!
             
             switch uploadFileResult {
             case .gone:
-                // TODO: Need the file uuid here too.
+                // TODO: Need the file uuid here too. And the uploadObjectTrackerId.
                 break
                 
-            case .success(let uploadResult):
+            case .success(let trackerId, let uploadResult):
                 fileUUID = uploadResult.fileUUID
+                uploadObjectTrackerId = trackerId
             }
             
             do {
-                if let uploadFileTracker = try UploadFileTracker.fetchSingleRow(db: db, where: fileUUID == UploadFileTracker.fileUUIDField.description),
+                // There can be more than one row in UploadFileTracker with the same fileUUID here because we can queue the same upload multiple times. Therefore, need to also search by uploadObjectTrackerId.
+                if let uploadFileTracker = try UploadFileTracker.fetchSingleRow(db: db, where: fileUUID == UploadFileTracker.fileUUIDField.description &&
+                    uploadObjectTrackerId == UploadFileTracker.uploadObjectTrackerIdField.description),
                     let url = uploadFileTracker.localURL {
                     if uploadFileTracker.uploadCopy {
                         try FileManager.default.removeItem(at: url)
@@ -60,7 +66,7 @@ extension SyncServer: ServerAPIDelegate {
             } catch let error {
                 delegate.error(self, error: error)
             }
-        
+
             delegate.uploadCompleted(self, result: uploadFileResult)
         }
     }
